AbstractWrapper Class

Abstract class representing a local dataset object in a Vitessce dataset.

Details

Abstract dataset object wrapper class

Methods

Public methods

AbstractWrapper$create_response_json()

AbstractWrapper$get_cells()

AbstractWrapper$get_cell_sets()

AbstractWrapper$get_raster()

AbstractWrapper$get_molecules()

AbstractWrapper$get_neighborhoods()

AbstractWrapper$get_expression_matrix()

AbstractWrapper$get_data()

AbstractWrapper$get_url()

AbstractWrapper$get_route()

AbstractWrapper$clone()

Method create_response_json()

Create a response function.

Usage

AbstractWrapper$create_response_json(data_list)

Arguments

data_list

The data to serve, as a list. Will be converted to JSON.

Returns

A new response function.

Method get_cells()

Get the routes and file definitions for the cells data type.

Usage

AbstractWrapper$get_cells(port, dataset_uid, obj_i)

Arguments

port

The port on which the web server is serving.

dataset_uid

The ID for this dataset.

obj_i

The index of this data object within the dataset.

Returns

NA or a list of routes and file_defs lists.

Method get_cell_sets()

Get the routes and file definitions for the cell sets data type.

Usage

AbstractWrapper$get_cell_sets(port, dataset_uid, obj_i)

Arguments

port

The port on which the web server is serving.

dataset_uid

The ID for this dataset.

obj_i

The index of this data object within the dataset.

Returns

NA or a list of routes and file_defs lists.

Method get_raster()

Get the routes and file definitions for the raster type.

Usage

AbstractWrapper$get_raster(port, dataset_uid, obj_i)

Arguments

port

The port on which the web server is serving.

dataset_uid

The ID for this dataset.

obj_i

The index of this data object within the dataset.

Returns

NA or a list of routes and file_defs lists.

Method get_molecules()

Get the routes and file definitions for the molecules data type.

Usage

AbstractWrapper$get_molecules(port, dataset_uid, obj_i)

Arguments

port

The port on which the web server is serving.

dataset_uid

The ID for this dataset.

obj_i

The index of this data object within the dataset.

Returns

NA or a list of routes and file_defs lists.

Method get_neighborhoods()

Get the routes and file definitions for the neighborhoods data type.

Usage

AbstractWrapper$get_neighborhoods(port, dataset_uid, obj_i)

Arguments

port

The port on which the web server is serving.

dataset_uid

The ID for this dataset.

obj_i

The index of this data object within the dataset.

Returns

NA or a list of routes and file_defs lists.

Method get_expression_matrix()

Get the routes and file definitions for the expression matrix data type.

Usage

AbstractWrapper$get_expression_matrix(port, dataset_uid, obj_i)

Arguments

port

The port on which the web server is serving.

dataset_uid

The ID for this dataset.

obj_i

The index of this data object within the dataset.

Returns

NA or a list of routes and file_defs lists.

Method get_data()

Get the routes and file definitions for a data type.

Usage

AbstractWrapper$get_data(data_type, port, dataset_uid, obj_i)

Arguments

data_type

A data type.

port

The port on which the web server is serving.

dataset_uid

The ID for this dataset.

obj_i

The index of this data object within the dataset.

Returns

NA or a list of routes and file_defs lists.

Method get_url()

Create a local web server URL for a dataset object data type.

Usage

AbstractWrapper$get_url(port, dataset_uid, obj_i, suffix)

Arguments

port

The port on which the web server is serving.

dataset_uid

The ID for this dataset.

obj_i

The index of this data object within the dataset.

suffix

A suffix for the URL.

Returns

A URL as a string.

Method get_route()

Create a web server route path for a dataset object data type.

Usage

AbstractWrapper$get_route(dataset_uid, obj_i, suffix)

Arguments

dataset_uid

The ID for this dataset.

obj_i

The index of this data object within the dataset.

suffix

A suffix for the URL.

Returns

A path as a string.

Method clone()

The objects of this class are cloneable with this method.

Usage

AbstractWrapper$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.