Enumeration of Coordination Type values

The CoordinationType list contains an enumeration of valid string constant values representing coordination types in the coordination space.

DATASET: The dataset to be used for a view.

EMBEDDING_TYPE: The type of embedding to be used in the embedding scatterplot view, for example "UMAP" or "t-SNE".

EMBEDDING_ZOOM: The zoom level of an embedding scatterplot view.

EMBEDDING_ROTATION: The rotation of an embedding scatterplot view.

EMBEDDING_TARGET_X: The x-axis center of an embedding scatterplot view.

EMBEDDING_TARGET_Y: The y-axis center of an embedding scatterplot view.

EMBEDDING_TARGET_Z: The z-axis center of an embedding scatterplot view.

SPATIAL_ZOOM: The zoom level of a spatial view.

SPATIAL_ROTATION: The rotation of a spatial view.

SPATIAL_TARGET_X: The x-coordinate of the center of a spatial view.

SPATIAL_TARGET_Y: The y-coordinate of the center of a spatial view.

SPATIAL_TARGET_Z: The z-coordinate of the center of a spatial view.

HEATMAP_ZOOM_X: The x-axis zoom level of a heatmap view.

HEATMAP_ZOOM_Y: The y-axis zoom level of a heatmap view.

HEATMAP_TARGET_X: The x-coordinate of the center of a heatmap view.

HEATMAP_TARGET_Y: The y-coordinate of the center of a heatmap view.

CELL_FILTER: A subset of cells to include after filtering.

CELL_HIGHLIGHT: A subset of cells to highlight.

CELL_SET_SELECTION: A subset of cell sets to select.

CELL_SET_HIGHLIGHT: A subset of cell sets to highlight.

CELL_SET_COLOR: A mapping from cell sets to colors.

GENE_FILTER: A subset of genes to include after filtering.

GENE_HIGHLIGHT: A subset of genes to highlight.

GENE_SELECTION: A subset of genes to select.

GENE_EXPRESSION_COLORMAP: The colormap to use for the gene expression scale.

GENE_EXPRESSION_COLORMAP_RANGE: The range of gene expression values to map.

CELL_COLOR_ENCODING: The color encoding to use for cell entities.

SPATIAL_LAYERS: Layer definitions for the spatial view.

GENOMIC_ZOOM: The zoom level of a higlass view.

GENOMIC_TARGET_X: The x-coordinate of the center of a higlass view.

GENOMIC_TARGET_Y: The y-coordinate of the center of a higlass view.

ADDITIONAL_CELL_SETS: User-defined cell sets.

CoordinationType

Format

An object of class list of length 32.