MultiImageWrapper Class

Class representing image objects in a Vitessce dataset.

Details

Image wrapper class

Super class

vitessce::AbstractWrapper -> MultiImageWrapper

Public fields

image_wrappers

The object to wrap.

use_physical_size_scaling

Whether or not to scale the image based on the physical size metadata stored in the file.

Methods

Public methods

MultiImageWrapper$new()

MultiImageWrapper$convert_and_save()

MultiImageWrapper$make_raster_routes()

MultiImageWrapper$make_raster_file_def_creator()

MultiImageWrapper$clone()

Inherited methods

vitessce::AbstractWrapper$auto_view_config()

vitessce::AbstractWrapper$get_file_defs()

vitessce::AbstractWrapper$get_out_dir()

vitessce::AbstractWrapper$get_out_dir_route()

vitessce::AbstractWrapper$get_route_str()

vitessce::AbstractWrapper$get_routes()

vitessce::AbstractWrapper$get_url()

Method new()

Create a wrapper around multiple image objects.

Usage

MultiImageWrapper$new(image_wrappers, use_physical_size_scaling = FALSE, ...)

Arguments

image_wrappers

A list of individual image wrapper objects.

use_physical_size_scaling

Whether or not to scale the image based on the physical size metadata stored in the file.

...

Parameters inherited from AbstractWrapper.

Returns

A new MultiImageWrapper object.

Method convert_and_save()

Create the JSON output files, web server routes, and file definition creators.

Usage

MultiImageWrapper$convert_and_save(dataset_uid, obj_i)

Arguments

dataset_uid

The ID for this dataset.

obj_i

The index of this data object within the dataset.

Method make_raster_routes()

Create a list representing the cells in the Seurat object.

Usage

MultiImageWrapper$make_raster_routes()

Returns

A list that can be converted to JSON.

Method make_raster_file_def_creator()

Make the file definition creator function for the cells data type.

Usage

MultiImageWrapper$make_raster_file_def_creator(dataset_uid, obj_i)

Arguments

dataset_uid

The ID for this dataset.

obj_i

The index of this data object within the dataset.

Returns

A file definition creator function which takes a base_url parameter.

Method clone()

The objects of this class are cloneable with this method.

Usage

MultiImageWrapper$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.