OmeTiffWrapper Class

Class representing an OME-TIFF file in a Vitessce dataset.

Details

OME-TIFF object wrapper class

Super class

vitessce::AbstractWrapper -> OmeTiffWrapper

Public fields

img_path

A local filepath to an OME-TIFF file.

img_url

A remote URL of an OME-TIFF file.

name

The display name for this OME-TIFF within Vitessce.

transformation_matrix

A column-major ordered matrix for transforming this image (see http://www.opengl-tutorial.org/beginners-tutorials/tutorial-3-matrices/#homogeneous-coordinates for more information).

is_bitmask

Whether or not this image is a bitmask.

is_remote

Whether or not this image is remote.

Methods

Public methods

OmeTiffWrapper$new()

OmeTiffWrapper$convert_and_save()

OmeTiffWrapper$make_raster_routes()

OmeTiffWrapper$make_image_def()

OmeTiffWrapper$make_raster_file_def_creator()

OmeTiffWrapper$clone()

Inherited methods

vitessce::AbstractWrapper$auto_view_config()

vitessce::AbstractWrapper$get_file_defs()

vitessce::AbstractWrapper$get_out_dir()

vitessce::AbstractWrapper$get_out_dir_route()

vitessce::AbstractWrapper$get_route_str()

vitessce::AbstractWrapper$get_routes()

vitessce::AbstractWrapper$get_url()

Method new()

Create a wrapper around multiple image objects.

Usage

OmeTiffWrapper$new(
  img_path = NA,
  img_url = NA,
  name = "",
  transformation_matrix = NA,
  is_bitmask = FALSE,
  ...
)

Arguments

img_path

A local filepath to an OME-TIFF file.

img_url

A remote URL of an OME-TIFF file.

name

The display name for this OME-TIFF within Vitessce.

transformation_matrix

A column-major ordered matrix for transforming this image (see http://www.opengl-tutorial.org/beginners-tutorials/tutorial-3-matrices/#homogeneous-coordinates for more information).

is_bitmask

Whether or not this image is a bitmask.

...

Parameters inherited from AbstractWrapper.

use_physical_size_scaling

Whether or not to scale the image based on the physical size metadata stored in the file.

Returns

A new OmeTiffWrapper object.

Method convert_and_save()

Create the web server routes and file definition creators.

Usage

OmeTiffWrapper$convert_and_save(dataset_uid, obj_i)

Arguments

dataset_uid

The ID for this dataset.

obj_i

The index of this data object within the dataset.

Method make_raster_routes()

Create a list representing the server routes.

Usage

OmeTiffWrapper$make_raster_routes(dataset_uid, obj_i)

Arguments

dataset_uid

The ID for this dataset.

obj_i

The index of this data object within the dataset.

Returns

A list of server route objects.

Method make_image_def()

Create an object representing a single image in a raster.json list of images.

Usage

OmeTiffWrapper$make_image_def(dataset_uid, obj_i, base_url)

Arguments

dataset_uid

The ID for this dataset.

obj_i

The index of this data object within the dataset.

base_url

The base URL for the server.

Returns

A list that can be converted to JSON.

Method make_raster_file_def_creator()

Make the file definition creator function for the raster data type.

Usage

OmeTiffWrapper$make_raster_file_def_creator(dataset_uid, obj_i)

Arguments

dataset_uid

The ID for this dataset.

obj_i

The index of this data object within the dataset.

Returns

A file definition creator function which takes a base_url parameter.

Method clone()

The objects of this class are cloneable with this method.

Usage

OmeTiffWrapper$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.