SeuratWrapper Class

Class representing a local Seurat object in a Vitessce dataset.

Details

Seurat object wrapper class

Super class

vitessce::AbstractWrapper -> SeuratWrapper

Public fields

obj

The object to wrap.

Methods

Public methods

SeuratWrapper$new()

SeuratWrapper$create_cells_list()

SeuratWrapper$create_cell_sets_list()

SeuratWrapper$get_cells()

SeuratWrapper$get_cell_sets()

SeuratWrapper$clone()

Inherited methods

vitessce::AbstractWrapper$create_response_json()

vitessce::AbstractWrapper$get_data()

vitessce::AbstractWrapper$get_expression_matrix()

vitessce::AbstractWrapper$get_molecules()

vitessce::AbstractWrapper$get_neighborhoods()

vitessce::AbstractWrapper$get_raster()

vitessce::AbstractWrapper$get_route()

vitessce::AbstractWrapper$get_url()

Method new()

Create a wrapper around a Seurat object.

Usage

SeuratWrapper$new(obj)

Arguments

obj

The object to wrap.

Returns

A new SeuratWrapper object.

Method create_cells_list()

Create a list representing the cells in the Seurat object.

Usage

SeuratWrapper$create_cells_list()

Returns

A list that can be converted to JSON.

Method create_cell_sets_list()

Create a list representing the cluster assignments in the Seurat object.

Usage

SeuratWrapper$create_cell_sets_list()

Returns

A list that can be converted to JSON.

Method get_cells()

Get the routes and file definitions for the cells data type.

Usage

SeuratWrapper$get_cells(port, dataset_uid, obj_i)

Arguments

port

The port on which the web server is serving.

dataset_uid

The ID for this dataset.

obj_i

The index of this data object within the dataset.

Returns

A list of routes and file_defs lists.

Method get_cell_sets()

Get the routes and file definitions for the cell sets data type.

Usage

SeuratWrapper$get_cell_sets(port, dataset_uid, obj_i)

Arguments

port

The port on which the web server is serving.

dataset_uid

The ID for this dataset.

obj_i

The index of this data object within the dataset.

Returns

A list of routes and file_defs lists.

Method clone()

The objects of this class are cloneable with this method.

Usage

SeuratWrapper$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.