Class representing a local web server to serve dataset objects.

Class representing a local web server to serve dataset objects.

Class representing a local web server to serve dataset objects.

Public fields

num_obj

The number of times the on_obj callback has been called.

Methods

Public methods

VitessceConfigServer$new()

VitessceConfigServer$create_routes()

VitessceConfigServer$run()

VitessceConfigServer$clone()

Method new()

Create a new server wrapper object.

Usage

VitessceConfigServer$new(port)

Arguments

port

The server port.

Returns

A new VitessceConfigServer object.

Method create_routes()

Set up the server routes.

Usage

VitessceConfigServer$create_routes(routes)

Arguments

routes

A list of VitessceConfigServerStaticRoute objects.

Method run()

Run the local server on the specified port.

Usage

VitessceConfigServer$run()

Method clone()

The objects of this class are cloneable with this method.

Usage

VitessceConfigServer$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.