VitessceConfigView Class

Class representing a view in a Vitessce view config.

Details

View in a VitessceConfig

Methods

Public methods

VitessceConfigView$new()

VitessceConfigView$use_coordination()

VitessceConfigView$set_xywh()

VitessceConfigView$to_list()

VitessceConfigView$clone()

Method new()

Create a new view object.

Usage

VitessceConfigView$new(component, coordination_scopes, x, y, w, h)

Arguments

component

A component name.

coordination_scopes

A list of the coordination scope mappings for each coordination type.

x

The x-coordinate of the view in the layout.

y

The y-coordinate of the view in the layout.

w

The width of the view in the layout.

h

The height of the view in the layout.

Returns

A new VitessceConfigView object.

Method use_coordination()

Link this view to existing coordination scope(s).

Usage

VitessceConfigView$use_coordination(c_scopes)

Arguments

c_scopes

A list of VitessceConfigCoordinationScope objects to use.

Returns

Invisible self, to allow chaining.

Method set_xywh()

Set the dimensions of the view.

Usage

VitessceConfigView$set_xywh(x, y, w, h)

Arguments

x

The x-coordinate of the view in the layout.

y

The y-coordinate of the view in the layout.

w

The width of the view in the layout.

h

The height of the view in the layout.

Method to_list()

Convert the object to an R list. Helpful when converting the config to JSON.

Usage

VitessceConfigView$to_list()

Returns

A list that can be serialized to JSON using rjson.

Method clone()

The objects of this class are cloneable with this method.

Usage

VitessceConfigView$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.