Vitessce Widget

This function creates a new Vitessce htmlwidget. A Vitessce widget is defined by a config which specifies dataset(s), a layout of views, and linked view connections. A config object can be passed to the widget using the config parameter.

vitessce_widget(
  config,
  theme = "dark",
  width = NULL,
  height = NULL,
  port = NA,
  base_url = NA,
  serve = TRUE,
  element_id = NULL
)

Arguments

config
A view config as a VitessceConfig object.

theme
The theme of the widget, either "dark" or "light". Optional. By default, "dark".

width
The width of the widget as a number or CSS string. Optional.

height
The height of the widget as a number or CSS string. Optional.

port
The port for the local web server (which serves local dataset objects to the widget). Optional. By default, uses open port between 8000 and 9000.

base_url
The base URL for the web server. Optional. By default, creates a localhost URL which includes the port.

serve
Should local data be served by running a local web server with R plumber? By default, TRUE.

element_id
An element ID. Optional.

Details

We do not recommend calling this function directly. Instead, we recommend calling the widget() method on the VitessceConfig instance.

Examples

vc <- VitessceConfig$new("My config")
vc$widget()